# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  docs.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: integer
    type: object
...
  docs.StandardResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: integer
    type: object
...
host: router.yeying.pub
info:
  contact: {}
  description: Router public/admin API (JWT/UCAN) under /api/v1.
  title: Router API
  version: v1
paths:
  /api/v1/admin/billing/currencies:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List billing currencies (admin)
      tags:
      - admin
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create billing currency (root)
      tags:
      - admin
  /api/v1/admin/billing/currencies/{code}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Currency code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update billing currency (root)
      tags:
      - admin
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"